Rename cityNameSubmitButtonClickHandler to submitHandler
