+++ src/client/index.js
@@ -1,10 +1,10 @@
 import { handleSubmit } from "./js/formHandler";
 import { isValueEmpty } from "./js/isValueEmpty";
-import { cityNameSubmitButtonClickHandler } from "./js/cityNameSubmitButtonClickHandler";
+import { submitHandler } from "./js/submitHandler";
 
 import "./styles/style.scss";
 
 const handleSubmitButton = document.getElementById("handleSubmit");
-handleSubmitButton.addEventListener("click", cityNameSubmitButtonClickHandler);
+handleSubmitButton.addEventListener("click", submitHandler);
 
-export { handleSubmit, isValueEmpty, cityNameSubmitButtonClickHandler };
+export { handleSubmit, isValueEmpty, submitHandler };
